requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asn1crypto==1.5.1
boto3==1.40.66
botocore==1.40.66
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
Faker==37.12.0
fastapi==0.121.0
filelock==3.20.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
jmespath==1.0.1
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
platformdirs==4.5.0
pluggy==1.6.0
pyarrow==22.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==4.0.0
sortedcontainers==2.4.0
starlette==0.49.3
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
